nvrs-agents/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
autogen-agentchat==0.2.37
certifi==2024.8.30
charset-normalizer==3.4.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
FLAML==2.3.1
grpcio==1.67.0
grpcio-tools==1.67.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
joblib==1.4.2
lightgbm==4.5.0
numpy==1.26.4
openai==1.52.2
packaging==24.1
pandas==2.2.3
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.9.11
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xgboost==2.1.2
